push predicate information through views to sources
